Preliminary Logistic regression classification using liblinear.
Bindings seems to work. However, some work is still needed:

    * Manage the special case when number of classes is 2. Some
      garbage is output to clf.coef_.

    * Implement weights.

    * Will we use bias > 0 ? (see liblinear's README)

    * Doc.

    * More tests.

    * Implement grouped classes.

From: Fabian Pedregosa <fabian.pedregosa@inria.fr>

git-svn-id: https://scikit-learn.svn.sourceforge.net/svnroot/scikit-learn/trunk@649 22fbfee3-77ab-4535-9bad-27d1bd3bc7d8

+++ scikits/learn/src/libsvm_helper.c
@@ -126,7 +126,7 @@ struct svm_problem * set_problem(char *X,char *Y, npy_intp *dims)
     struct svm_problem *problem;
     problem = (struct svm_problem *) malloc(sizeof(struct svm_problem));
     if (problem == NULL) return NULL;
-    problem->l = dims[0];
+    problem->l = (int) dims[0];
     problem->y = (double *) Y;
     problem->x = dense_to_sparse((double *) X, dims);
     if (problem->x == NULL) { 
@@ -236,7 +236,7 @@ void copy_sv_coef(char *data, struct svm_model *model, npy_intp *strides)
 
 void copy_rho(char *data, struct svm_model *model, npy_intp *dims)
 {
-    memcpy(data, model->rho, dims[0]*sizeof(double));
+    memcpy(data, model->rho, dims[0] * sizeof(double));
 }
 
 /*
